DecodeMainExe.py
DecodeMainExe.py is a de-obfuscation and decryption console-line Python3 program. 
It's specifically designed and intended to function as a decoder for Kovter infected .EXE executables, as the algorithm is especially designed for that, but the code is of course modifyable and can easily be altered according to individual wants and needs, or to add additional capabilities.